soc/intel/common/xhci: Fix building for 64bit
Tested with clang on prodrive/hermes: Boots to payload

Reviewed-on: https://review.coreboot.org/c/coreboot/+/69234
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/soc/intel/common/block/xhci/xhci.c
@@ -27,7 +27,7 @@ static uint8_t *xhci_mem_base(void)
 	if (mem_base == 0 || mem_base == 0xffffffff)
 		return 0;
 
-	return (uint8_t *)(mem_base & ~PCI_BASE_ADDRESS_MEM_ATTR_MASK);
+	return (uint8_t *)(uintptr_t)(mem_base & ~PCI_BASE_ADDRESS_MEM_ATTR_MASK);
 }
 
 void xhci_host_reset(void)
